+++ qase-javascript-commons/src/models/index.ts
@@ -1,9 +1,11 @@
-export { TestResultType, Relation, Suite, SuiteData } from './test-result';
+export { TestResultType } from './test-result';
+export type { Relation, Suite, SuiteData } from './test-result';
 export { TestExecution, TestStatusEnum } from './test-execution';
 export { TestStepType, StepType } from './test-step';
 export { StepStatusEnum } from './step-execution';
-export { Attachment } from './attachment';
-export { Report } from './report';
+export type { Attachment } from './attachment';
+export type { Report } from './report';
 export { CompoundError } from './error';
-export { ConfigurationGroup, ConfigurationItem, ConfigurationGroupResponse } from './configuration';
-export { ExternalLinkType, TestOpsOptionsType, TestOpsApiType, TestOpsRunType, TestOpsPlanType, TestOpsBatchType, TestOpsConfigurationType, TestOpsConfigurationValueType, TestOpsExternalLinkType } from './config/TestOpsOptionsType';
+export type { ConfigurationGroup, ConfigurationItem, ConfigurationGroupResponse } from './configuration';
+export { ExternalLinkType } from './config/TestOpsOptionsType';
+export type { TestOpsOptionsType, TestOpsApiType, TestOpsRunType, TestOpsPlanType, TestOpsBatchType, TestOpsConfigurationType, TestOpsConfigurationValueType, TestOpsExternalLinkType } from './config/TestOpsOptionsType';