+++ qase-vitest/jest.config.js
@@ -9,7 +9,6 @@ module.exports = {
       tsconfig: {
         skipLibCheck: true,
       },
-      isolatedModules: true,
     }],
   },
   collectCoverageFrom: [